chore: Rename Event classes
